requirements.txt
lxml==4.5.2
redis==3.5.3
pyquery==1.4.1
requests==2.24.0
python-dotenv==0.14.0
